+++ CHANGES.md
@@ -1,3 +1,11 @@
+v0.3 (5/2022)
+=============
+- sample rate is a mandatory argument
+- user can input sample rate as 192k instead of 192000
+- user can specify de-emphasis time constant with -d
+- adjusted de-emphasis to be 50 μs by default
+- adjusted audio low-pass from 15 kHz to 16.5 kHz
+
 v0.2 (1/2022)
 =============
 - correct stereo polarity